+++ NEWS.md
@@ -1,5 +1,11 @@
 # smoof 1.6.0.4
 
+## New features
+
+* Added support for the extended bi-objective BBOB functions (FIDs 56-92) in makeBiObjBBOBFunction
+* Added extra helper functions to extract all problem instance-specific data from MPM2 generator
+* Added R-based evaluation environment for MPM2 generator which greatly accelerates evaluation in multi-objective settings
+
 ## Bugfixes 
 
 * Fixed bug in getLoggedValues when logging of x-values was set to FALSE in addLoggingWrapper